sample-tractor-repair-assistant-agent/requirements.txt
oracledb
langchain
langchain-community
langchain-openai
langchain-oracledb
langchain-text-splitters
langgraph
tiktoken
python-dotenv
tavily-python
lark
langchain-tavily
pydantic
python-dotenv
fastapi
uvicorn